pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "lustro"
version = "0.1.0"
description = "Survey and illuminate the AI/tech landscape."
requires-python = ">=3.10"
dependencies = [
  "feedparser>=6",
  "requests>=2.28",
  "trafilatura>=1.6",
  "pyyaml>=6",
  "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
digest = [
  "openai>=1.0",
  "httpx>=0.24",
]

[project.scripts]
lustro = "lustro.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lustro"]

[tool.hatch.build.targets.wheel.force-include]
"src/lustro/sources/default.yaml" = "lustro/sources/default.yaml"

[tool.hatch.build.targets.sdist]
include = [
  "src/lustro/**/*.py",
  "src/lustro/sources/default.yaml",
  "tests/**/*.py",
  "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
